added Expection handling V1

+++ 3220FinalProject/adminview.cpp
@@ -1,6 +1,8 @@
 #include "adminview.h"
 #include "ui_adminview.h"
 #include "additemwindow.h"
+#include "error.h"
+#include "iostream"
 
 AdminView::AdminView(QWidget *parent) :
     QDialog(parent),
@@ -33,18 +35,79 @@ void AdminView::on_pushButton_LoadTable_clicked()
 
 void AdminView::on_pushButton_Delete_clicked()
 {
+
+    int flag = 1;
     //Need to add error handling
     QString id = ui->lineEdit_DeleteId->text();
+
+    AdminError AE(data);
+
+    try{
+        //make sure is not empty
+        if(AE.isEmpty(id) == 0){
+            throw 0;
+        }
+    }
+        catch(int &d){
+            flag = 0;
+            qDebug() << "Item ID is empty, not able to delete" << endl;
+          //  throw "Item ID is empty, not able to delete";
+
+            }
+
+    try{
+        //make sure it is all integer
+        if(AE.typeCheck(id) == 0){
+            throw 0;
+        }
+    }
+
+        catch(int &d){
+        flag = 0;
+
+            qDebug() << "Id is not Integer" << endl;
+          //  throw "Id is not Integer";
+
+        }
+
     int ItemId = id.toInt();
+
+    try{
+        if(AE.deleteMatch(ItemId) == 0){
+            throw 0;
+        }
+    }
+        catch(int &d){
+        flag = 0;
+
+            qDebug() << "Unable to delete" << endl;
+         //   throw "Unable to delete";
+
+        }
+    catch(...){
+        flag = 0;
+
+        qDebug() << "catch all" << endl;
+
+    }
+
+
     cout<<"correct ID is" << ItemId;
-    QSqlQuery query;
-    query.prepare("DELETE from Item Where ItemID = :ItemId");
-    query.bindValue(":ItemId",ItemId);
-    query.exec();
+    if(flag == 1){
+        QSqlQuery query;
+        query.prepare("DELETE from Item Where ItemID = :ItemId");
+        query.bindValue(":ItemId",ItemId);
+        query.exec();
+
+    }
+
+
+
 }
 
 void AdminView::on_pushButton_AddItem_clicked()
 {
+
 
     AddItemWindow aiw;
     aiw.data = data;
